requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-middleware==1.5.0
django-extensions==3.1.0
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.9.1
PyJWT==2.0.0
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
django-filter==2.4.0
django-autoslug==1.9.8
celery[redis]==5.0.5
channels==3.0.3
channels-redis==3.2.0